Add dev env setup
