+++ apps/search-poc/package.json
@@ -52,6 +52,7 @@
     "@types/node": "12.6.8",
     "@types/react": "16.8.23",
     "@types/react-dom": "16.8.5",
+    "@webpack-cli/serve": "^0.1.8",
     "abort-controller": "^3.0.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-dynamic-import-webpack": "^1.1.0",
@@ -61,6 +62,8 @@
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "style-loader": "^1.0.0",
     "terser-webpack-plugin": "^1.4.1",
-    "typescript": "3.5.3"
+    "typescript": "3.5.3",
+    "webpack-cli": "^3.3.7",
+    "webpack-dev-server": "^3.8.0"
   }
 }
