+++ apps/search-poc/src/web-components/resource-details/webpack.config.prod.js
@@ -1,61 +1,15 @@
 
-const path = require('path');
 const webpack = require('webpack');
-
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
 
+const commonConfig = require('./webpack.config.common');
+
 
 module.exports = {
+  ...commonConfig,
   mode: 'production',
-  entry: './src/web-components/resource-details',
-  output: {
-      path: path.resolve(__dirname, '../../../dist-wc/'),
-      filename: 'resource-details.bundle.min.js'
-  },
-  resolve: {
-      extensions: ['.ts', '.tsx', '.js', '.json']
-  },
-  module: {
-      rules: [{
-        test:/\.css$/,
-        use: [{
-          loader: 'style-loader',
-          options: {
-            injectType: 'singletonStyleTag',
-            insert: function (styleElement) {
-              window.getResourceDetailsStyleEl = function() {
-                return styleElement;
-              }
-            },
-          },
-        },
-          'css-loader',
-        ]
-    }, {
-          test: /\.(ts|js)x?$/,
-          exclude: /node_modules/,
-          loader: 'babel-loader',
-          options: {
-            presets: [
-              '@babel/preset-env',
-              "@babel/preset-react",
-              "@babel/preset-typescript"
-            ],
-            plugins: [
-              ['import', { libraryName: 'antd', libraryDirectory: 'es', style: 'css' }],
-              "@babel/proposal-class-properties",
-              "@babel/proposal-object-rest-spread",
-            ]
-          }
-      }],
-  },
   plugins: [
-    new HtmlWebpackPlugin({
-      template: './src/web-components/resource-details/template.html',
-      inject: 'body',
-    }),
     new webpack.optimize.LimitChunkCountPlugin({ maxChunks: 1 })
   ],
   optimization: {
